about_env Feat : Ajout de la fonction qui compter le nombre de caractere dans le shell->exit_code about create_tab Fix : Meilleur gestion du manage_count qui comptais a l'infinie si (\ = double quote) $?hassmou\ | ahsh'hassmou'$USER\ AHAH donc si il y avait '' puis directement \. Je n'avais pas gerer le cas ou il y aurait les 2 qui s'enchainent vu qu'il y avait l'index qui faisait +1 si un quote etait finit.ainsi il ne voyait pas l'autre qui apparaissait.

+++ Parsing/Parser/change_fd.c
@@ -6,13 +6,13 @@
 /*   By: hassmou <hassmou@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/29 19:16:18 by hassmou           #+#    #+#             */
-/*   Updated: 2026/07/15 17:18:40 by hassmou          ###   ########.fr       */
+/*   Updated: 2026/07/18 17:29:35 by hassmou          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parsing.h"
 
-int	manage_fd(t_tokens **tokens, t_cmd *cmd, int redir)
+int	manage_fd(t_tokens **tokens, t_cmd *cmd, size_t redir)
 {
 	if (((*tokens)->type == redir) && (redir == REDIR_IN))
 	{
